[fe#15] refactor: header 컴포넌트 분리
